Add support for FlexUnit to GradleFx
Add support for compiling and executing FlexUnit tests from the GradleFx
plug-in. The Mxmlc compiler task class is now split in 2 as we need to
call it as a normal application/module compilation as well as compiling
a test runner for FlexUnit.

Includes various small refactorings to clean up the code and
simplify things. Add comments to the major classes to make it more
obvious the role they play in the plug-in.
things.